fix warnings and an actual error in vgmwrite.h (buffer too short)
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5236 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/state.c
@@ -89,7 +89,7 @@ static int ss_current_tag;
 
 static unsigned char *ss_dump_array;
 static mame_file *ss_dump_file;
-static unsigned int ss_dump_size;
+static size_t ss_dump_size;
 
 
 static UINT32 ss_get_signature(void)
